Rename classes and refactor custom security services
