+++ src/main/java/com/alkemy/ong/application/service/LoginUserUseCaseService.java
@@ -3,7 +3,7 @@ package com.alkemy.ong.application.service;
 import com.alkemy.ong.application.exception.ErrorMessage;
 import com.alkemy.ong.application.exception.ObjectNotFound;
 import com.alkemy.ong.application.repository.IUserRepository;
-import com.alkemy.ong.application.service.delegate.IAuthenticateUser;
+import com.alkemy.ong.application.service.delegate.IAuthenticationManager;
 import com.alkemy.ong.application.service.usecase.ILoginUserUseCase;
 import com.alkemy.ong.domain.User;
 import java.util.Optional;
@@ -13,11 +13,11 @@ import lombok.AllArgsConstructor;
 public class LoginUserUseCaseService implements ILoginUserUseCase {
 
   private IUserRepository userRepository;
-  private IAuthenticateUser authenticateUser;
+  private IAuthenticationManager authenticationManager;
 
   @Override
   public User login(User user) {
-    authenticateUser.authenticate(user);
+    authenticationManager.authenticate(user);
     return findBy(user.getEmail());
   }
