+++ src/main/java/com/alkemy/ong/application/service/delegate/IAuthenticationManager.java
@@ -2,7 +2,7 @@ package com.alkemy.ong.application.service.delegate;
 
 import com.alkemy.ong.domain.User;
 
-public interface IAuthenticateUser {
+public interface IAuthenticationManager {
 
   void authenticate(User user);
