+++ src/main/java/com/alkemy/ong/infrastructure/config/spring/SpringBeanConfiguration.java
@@ -21,7 +21,8 @@ import com.alkemy.ong.application.service.GetOrganizationUseCaseService;
 import com.alkemy.ong.application.service.ListSlideUseCaseService;
 import com.alkemy.ong.application.service.LoginUserUseCaseService;
 import com.alkemy.ong.application.service.UpdateCategoryUserCaseService;
-import com.alkemy.ong.application.service.delegate.IAuthenticateUser;
+import com.alkemy.ong.application.service.delegate.IAuthenticationManager;
+import com.alkemy.ong.application.service.delegate.IOperationAllowed;
 import com.alkemy.ong.application.service.usecase.ICreateUserUseCase;
 import com.alkemy.ong.application.service.usecase.IDeleteCategoryUseCase;
 import com.alkemy.ong.application.service.usecase.IDeleteCommentUseCase;
@@ -33,7 +34,6 @@ import com.alkemy.ong.application.service.usecase.IDeleteUserUseCase;
 import com.alkemy.ong.application.service.usecase.IGetOrganizationUseCase;
 import com.alkemy.ong.application.service.usecase.IListSlideUseCase;
 import com.alkemy.ong.application.service.usecase.ILoginUserUseCase;
-import com.alkemy.ong.application.service.delegate.IOperationAllowed;
 import com.alkemy.ong.application.service.usecase.IUpdateCategoryUseCase;
 import com.alkemy.ong.infrastructure.database.repository.CategoryRepository;
 import com.alkemy.ong.infrastructure.database.repository.SlideRepository;
@@ -105,8 +105,8 @@ public class SpringBeanConfiguration {
 
   @Bean
   public ILoginUserUseCase loginUser(IUserRepository userRepository,
-      IAuthenticateUser authenticateUser) {
-    return new LoginUserUseCaseService(userRepository, authenticateUser);
+      IAuthenticationManager authenticationManager) {
+    return new LoginUserUseCaseService(userRepository, authenticationManager);
   }
 
 }
